% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_economy.R
\name{simulate_economy}
\alias{simulate_economy}
\title{Simulate an Monetary Policy}
\usage{
simulate_economy(
  periods = 50,
  gamma = 0.75,
  alpha = 0.8,
  rule_pi = 1.5,
  rule_y = 0.5,
  delta = 1,
  adaptive = 1,
  cost_push_shock_persistance = 0.5,
  demand_shock_persistance = 0.5,
  monetary_shock_persistance = 0.5,
  financial_shock_persistance = 0.5,
  demand_shocks = c(),
  cost_push_shocks = c(),
  monetary_shocks = c(),
  financial_shocks = c(),
  demand_shock_magnitude = 1,
  cost_push_shock_magnitude = 1,
  monetary_shock_magnitude = 1,
  financial_shock_magnitude = 1
)
}
\arguments{
\item{periods}{Number of periods to simulate (Default is 50)}

\item{gamma}{Parameters of the Phillips curve}

\item{alpha}{Parameters of the IS curve}

\item{rule_pi}{Parameters of the interest rate rule; (Response to inflation (relative to target))}

\item{rule_y}{Parameters of the interest rate rule; (Response to output gap )}

\item{delta}{Financial Markets frictions}

\item{adaptive}{Weight on past inflation (1=adaptive expectations; 0=rational expectations)}

\item{financial_shock_magnitude}{}
}
\value{
a named list
}
\description{
Simulate an Monetary Policy
}
