% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_economy.R
\name{plot.simulate_economy}
\alias{plot.simulate_economy}
\title{Plot Economy}
\usage{
\method{plot}{simulate_economy}(x, ...)
}
\arguments{
\item{x}{- an object of class simulate_economy}

\item{...}{used for future expansions}
}
\value{
a plot
}
\description{
plot generic for simulate_economy class
}
\examples{
econ1 <- simulate_economy(
  cost_push_shock_persistance = 0.5,
  demand_shock_persistance = 0.75,
  monetary_shock_persistance = 0.5,
  financial_shock_persistance = 0.5,
  demand_shocks = c(3, 10),
  cost_push_shocks = c(5, 15),
  monetary_shocks = c(7),
  financial_shocks = c(9),
  demand_shock_magnitude = c(-2, 1),
  cost_push_shock_magnitude = c(1.5, -0.5),
  monetary_shock_magnitude = c(1),
  financial_shock_magnitude = c(0.5)
)

plot(econ1)

}
