% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_economy.R
\name{plot.simulate_economy}
\alias{plot.simulate_economy}
\title{PLot Economy}
\usage{
\method{plot}{simulate_economy}(x, ...)
}
\arguments{
\item{x}{- an object of class simulate_economy}

\item{...}{used for future expansions}
}
\value{
a plot
}
\description{
plot generic for simulate_economy class
}
\examples{
econ1 <- simulate_economy(
             adaptive = 0,
             demand_shocks = c(3),
             cost_push_shocks = c(3),
             monetary_shocks = c(3),
             financial_shocks = c(5),
             cost_push_shock_magnitude = 4)

plot(econ1)

}
